Allow segments to restore globals and schemas
Currently pg_upgrade upgrade the coordinator and then copies the master
data directory to the primaries to upgrade the segments. Because of this
there were a number of special code paths for primaries. This was done
because pg_dump's initial implementation was too slow to run on
segments. Now that pg_dump is fast enough to run on segments,
coordinator and segment code paths can be consolidated and simplified.