Ignore pg_aovisimap name differences
The old and new cluster aux relation name may be different. Comment from
pg_dump explains why.

Pg_dump adjusts the names of all pg_aoseg objects for the given table to
match what they will be in the new cluster, using the OID of the owning
table. In many cases these will be the same as the ones in the old
cluster, but not always. Some operations can cause the old pg_aoseg
object names not to match its owner's OID, but the new cluster will be
using the correct name, and it's the new cluster's name that we have to
use.

Co-authored-by: Brent Doil <bdoil@vmware.com>