pg_upgrade: preserve gp_fastsequence to prevent duplicate ctids
The gp_fastsequence catalog table must be preserved when upgrading to a
newer version. Not preserving this table will likely cause duplicate
ctids in AO tables when normal cluster operations resume after an
upgrade.

How duplicate ctids can occur on AO tables:
1. gp_fastsequence is initialized to 0 for a new AO table.
2. gp_fastsequence increments positively in mulitples of 100.
3. ctids are assigned based on gp_fastsequence values.
4. When the an AO table's fastsequence value is 0 or missing,
   fastsequence is initialized to table's tuple count.

When a tuple is inserted it consumes a gp_fastsequence value to produce
a ctid to index the data. If a fastsequence value is reused, a duplicate
ctid will occur. This will result in indexes returning wrong results.

When an AO table's fastsequence value is 0 or missing, gpdb's default
behavior in an attempt to fix the fastsequence value is to set it to the
table's tuple count on the next insert. However this not reliable since
fastsequence reserves numbers in increments of 100 at a time for
performance purposes. This means that even if only 1 value is inserted
into the table, fastsequence will increment by 100. Under normal cluster
operation, fastsequence increments at a much faster rate than tuple
count and setting it to a table's tuple count can still result in a
duplicate ctid because there will be rows with ctids based on
fastsequence values much higher than its table's tuple count.