pg_upgrade: add support func to check if view contains removed functions
Views that use removed functions will cause upgrade to fail. This
happens during metadata restore on the target cluster because pg_restore
will error trying to create a view using functions that do not exist
anymore. This is not ideal as we could be many hours into upgrade before
a view using a removed function causes pg_upgrade to fail. In order to
detect views with removed functions. A query_tree_walker is used to walk
the nodes of all views. Each node is then checked to see if it is a
function. If it is, we then check if the oid of the function matches one
that is not present in GPDB7.

Currently we only notify end users of views that contain a removed
function. We log what specific part of the view triggered the check for
a removed function.

The following functions were excluded from being checked.
 - Overloaded functions that still existed but had their function
   signature changed.
 - Functions that were related to removed types. It is assumed that the
   removed types check will catch these functions.
 - Functions moved to another schema (pg_resgroup_move_query).