Support restore point based snapshot isolation for hot standby
Currently, the transaction isolation supported on hot standby has a number of
key limitations:

1. It is possible that different standby segments have replayed to different WAL
  records at a given time (for example, seg0 replays an INSERT, but seg1 hasn't).
  So the query on standby could easily break ACID.
2. As one of the query conflict types for hot standby, the snapshot conflict in
  GPDB is not taking distributed snapshot into account, causing potential
  correctness issues. See isolation2/hot_standby/queryconflict test for example.
3. The hot standby is not able to view the last one-phase commit (1PC) visible,
  even if the standby segments has replayed them. This is because since 1PC does
  not have any WAL on QD, the standby QD won't advance its latestCompletedGxid,
  so its distributed snapshot horizon will not include the last 1PC - it would
  view them not-yet-started or at best, still-in-progress.
4. The primary segments truncate the distributed log that is no longer used by
  any distributed snapshots *on the primary nodes*. However, it does not
  account for distributed snapshot on the standby nodes. As a result, the
  standby which has truncated distributed log that it still needs might fail to
  do dtx visibility checks, and give wrong results.

The key problem here is that, we are using distributed snapshot on hot standby.
But distributed snapshot as a way to describe the current transaction state in
a primary GPDB cluster, cannot do the same for hot standby. The distributed
snapshot taken on the standby QD essentially describes the transaction state
that a primary QD sees on the primary cluster (plus whatever read-only queries
running on the hot standby). But the state on the standby cluster can be totally
different, due to different pace of WAL replay on each segment.

We now use a restore point based approach to address all the above issues. The
overaching idea is that, we move away from distributed snapshot on hot standby,
and introduce a new "snapshot mode", namely "Restore Point Based Snapshot".
Every standby coordinator/segment will just use local snapshot, which is taken
at a certain restore point. At each restore point, it is by design guaranteed
that a distributed transaction is either in-progress on ALL segments, or
committed on ALL segments. Therefore, there is no need for distribution
coordination, hence no need for distributed snapshot.

This new mode breaks a couple of ABI, most notably in SnapshotData and
DtxContextInfo, which keeps track of "restore point name" needed by our new
mode. We deem it OK as they are internal structs.

The basic mechanism is:
* The primary creates restore points (RP) using gp_create_restore_point(),
  preferably in some periodic fashion (e.g. every 5 minutes).
* The standby coordinator/segment replays the RP WALs. Upon replaying each RP
  WAL, every standby coordinator/segment exports a snapshot, and remembers the
  RP name.
* When a query comes, the standby coordinator chooses an RP to use (either
  specified by the user, or just using the latest RP). Then, it imports a
  snapshot corresponding to that RP which will be used for the entirety of the
  transaction.
* When the standby coordinator dispatches the query, it includes the RP name in
  the dispatch message. Similar to what happened on the coordinator, the standby
  segmetns will import a snapshot corresponding to that RP.
With this design, we no longer support read-committed isolation. It will always
be repeatable-read, since the snapshot doesn't change in a transaction.

We introduce 3 GUCs to control the behavior:

1. gp_hot_standby_snapshot_mode: set the snapshot mode on standby. Only
   superuser can set it. The options are:
   * 'restorepoint', which is the default. This is what we would like to
      encourage the user to use for consistent results.
   * 'inconsistent', which turns off the RP-based approach. As mentioned
      initially, this would give possibly inconsistent result.
2. gp_hot_standby_snapshot_restore_point_name: set the restore point name to
   be used on the standby. Session-level GUC.
3. gp_max_restore_point_snapshots: set the max restore points the server keeps
   track of. The default is 1024.

Here is a simple example usage:

```
gpconfig -c gp_hot_standby_snapshot_mode -v restorepoint;
gpstop -ar;

set gp_hot_standby_snapshot_restore_point_name = 'rp1';
select * from t;
```

Reviewed-by: Soumyadeep Chakraborty soumyadeep2007@gmail.com
Reviewed-by: Jimmy Yih jyih@vmware.com