Add GUC to load gp_segment_configuration from a file
Sometimes gp_segment_configuration catalog might not reflect
the actual configuration of the cluster.

Consider the scenario where we have a secondary cluster which might be under
active recovery. This cluster will most likely be read only and its
gp_segment_configuration table will reflect information about the initial
cluster including hosts, ports and data directories which might be invalid.
For cases like these, we need the ability to make the gp_segment_configuration
reflect the correct information corresponding to the cluster.

This commit introduces a new GUC `gp_segment_configuration_file` which
specifies the path to a file that contains the correct segment
configuration information for the cluster. This file can then be read
to initialize gp_segment_configuration in-memory when creating the cdbgang.

Note that this file is created by the user and is different from the FTS file.
However their format is the the same.

We also skip FTS cycles while this guc is set because FTS
is not supposed to run on the cluster under active recovery
including during the time of promotion of the recovery cluster
until it's fully completed.

Here is an example of how the file looks like:
```
1 -1 p p n u 7000 coordinator_host coordinator_host /path/to/DataDir-1
2 0 p p s u 7002 segment0_host segment0_host /path/to/DataDir0
3 1 p p s u 7003 segment1_host segment1_host /path/to/DataDir1
4 2 p p s u 7004 segment1_host segment1_host /path/to/DataDir2

```

And the corresponding gp_segment_configuration entries:
```
 dbid | content | role | preferred_role | mode | status | port |    hostname    |    address     |       datadir

------+---------+------+----------------+------+--------+------+------------------+--------------------------------------
    1 |      -1 | p    | p              | n    | u      | 7000 |coordinator_host|coordinator_host| /path/to/DataDir-1
    2 |       0 | p    | p              | s    | u      | 7002 | segment0_host  | segment0_host  | /path/to/DataDir0
    3 |       1 | p    | p              | s    | u      | 7003 | segment1_host  | segment1_host  | /path/to/DataDir1
    4 |       2 | p    | p              | s    | u      | 7004 | segment2_host  | segment2_host  | /path/to/DataDir2

```