Support snapshot conflict on hot standby for ao/co tables and index
The snapshot conflict on hot standby requires the VACUUM running on primary node
to emit the latestRemovedXid among the rows being vacuumed. For ao/co tables,
however, their tuples do not have xid information stored with them. Ao/co vacuum
works in the way that a whole segfile is being compacted into a new segfile, and
the original segfile is being truncated. In this case, we need the
latestRemovedXid among the tuples in the truncated segfile. Unfortunately,
because of the way ao/co vacuum works, this xid will almost always be the xid of
the VACUUM transaction itself, which has higher chance to cause conflict
(compared to heap vacuum). A good way to think about it is that an ao/co VACUUM
can be described as an "INSERT (into new segfile)" followed by a "DELETE (the
original segfile)" and finally a "VACUUM (dead tuples in the original segfile)";
And the latestRemovedXid tracks the xid of the DELETE, which is just the same
one as the VACUUM itself. This is unless no tuple is being compacted in the
original segfile, which is very rare and we don't consider it for now.

However, in many cases the final truncation doesn't happen until later, when no
other transaction can see the deleted segfile. This decision is made by checking
the xmin of the corresponding AWAITING_DROP row in the aoseg table, which was
inserted by the initial VACUUM, against oldest xmin among running transactions.
Similarly, for snapshot conflict purpose, this is also the latestRemovedXid a
hot standby needs to check against its own oldest xmin among running
transactions. So to support snapshot conflict, we are now emiting this value in
WAL so that hot standby can detect the conflict. We re-use the existing
XLOG_HEAP2_CLEANUP_INFO record because it is not really necessary to introduce a
new WAL type when the content is exactly the same.

For ao/co index, the above latestRemovedXid is also equivalent to the same among
to-be-vacuumed index tuples. The key observation here is that we are truncating
ao segfile/vacuuming ao index tuples all in the batches of segfiles. So as long
as the standby processes the latestRemovedXid first, both ao/co table and index
conflict can be covered.

In practice, the conflict might be triggered in even earlier places during the
ao/co VACUUM path, most notably during scanning heap table (like aux tables
etc.) which might emits XLOG_HEAP2_CLEAN records. In addition, if one vacuums
ao/co table with AO_AUX_ONLY, that could also trigger conflict before the actual
segfile truncation. But in any cases, our new WAL record acts as the final guard
before doing the actual table/index vacuuming, which cancels any conflicting
standby query before they could see wrong results due to the table/index
vacuuming.

Thanks to Soumyadeep Chakraborty and Ashwin Agrawal for many discussions on the
solution ideas.