Try to fix flaky test dispatch
Diffs:
```
@@ -247,8 +247,13 @@
 \c dispatch_test_db
 -- verify failure will be reported
 SELECT 1 FROM gp_dist_random('gp_id');
-DETAIL:  Internal error: No motion listener port
-ERROR:  failed to acquire resources on one or more segments
+ ?column?
+----------
+        1
+        1
+        1
+(3 rows)
+
```

There's no pg log for this flake, the reason is *probably* some background
worker came in and dispatched something that triggered the fault on QE.
Although we knoew that it couldn't be FTS (as protected by a check around
sendQEDetails() in PostgresMain()), nor autovacuum (as the test itself is
running w/o autovacuum), but it could still be things like dtx recovery process.
So now using gp_inject_fault_infinite to try to fix the issue. It would be
generally better anyway.