Set synchronous_standby_names default value to "gp_walreceiver" on segment (#17574)
* Set the synchronous_standby_names default value to "gp_walreceiver" on segment.

If the default value of synchronous_standby_names is "*", then all replication
slot->priority will be larger than 0 which means sync mode.
For logical replication slot, the decoding normally from a source database, so
other database's transaction will not be processed, so not all wal logs are
processed and so the `flush` value could not updated as fast as sync mode. So it
may cause segment hang on commit SyncRepWaitForLSN waiting for logical
replication slot->flush to update.

With synchronous_standby_names set to "gp_walreceiver", logical replication slot
->priority will be set to 0 which means not sync mode which is reasonable.

This PR also add one check for the coordinator node that only when the priority is not 0, 
will the function `SyncRepGetCandidateStandbys` pick the walsender as one candidate.