Allocate dtx snapshot inProgressXidArray out of DistributedSnapshot_Reset
It is weird that DistributedSnapshot_Reset() which is supposed to reset
everything to 0 is responsible to allocate memory space for inProgressXidArray.
It is also inefficient that we allocate the space regardless of whether we need
distributed snapshot or not. The original commit 37d683e6d8e3 doesn't have much
information about that.

It should be safe to allocate the space in higher layer, where we have the
information about whether distributed snapshot is used or not. Same as before,
we won't malloc again if it is already allocated, so we should have less or
equal memory footprint than before.

Another goal of this change is that, we are planning to introduce a new snapshot
mode than the distributed snapshot (see #17561). The relative info of that mode
will be placed in the same memory space as distributed snapshot, as a union.
With this change, we can have a wholesale reset of the union space w/o worrying
about not initialize inProgressXidArray when it's needed.