Fix flakiness involving is_query_waiting_for_syncrep
The test like segwalrep/syncrep_throttle is flaky with diffs like:
```
@@ -38,7 +38,7 @@
 SELECT is_query_waiting_for_syncrep(50, 'INSERT INTO insert_throttle SELECT 1 FROM generate_series(1, 1000000);');
  is_query_waiting_for_syncrep
 ------------
- t
+ f
 (1 row)
```

The reason is that is_query_waiting_for_syncrep() is too fast to finish
checking even 50 times. A simple fix is to put an interval between each check.

An alternative fix is to wait until 'wal_sender_loop' is triggered before we
call is_query_waiting_for_syncrep(). But the current way is much simpler.