Fix Orca's exhaustive2 join order to consider random table's distribution (#17635)
Previously, commit 0c43cb5 added consideration of distribution specs to
improve dynamic partition elimination plan generation within the DPv2
(exhaustive2) transform. However, it didn't consider randomly
distributed tables when checking for distribution column overlap.

More specifically, it was looking for a particular join tree pattern and
ensuring that the join condition contained the partition tables
distribution columns so that no motion would need to be added (to
prevent generating an invalid alternative). Since randomly distributed
tables don't have a distribution column, the previous logic would always
think this pattern "matched" with a random table and that it was a good,
valid alternative. This alternative would be thrown away in the later
optimization step when it was determined to be invalid.

This commit adds an explicit check for randomly distributed tables
within DPv2, and generates only valid plans with partition selectors.

With this change, I see that on the 3TB tpcds suite with randomly
distributed tables we get some plan changes as follows:

17 - Extra PS, 52s -> 9s
25 - 2 Extra PS, 45s -> 7s
29 - Extra PS, 48s -> 24s
91 - Slight join order change, tiny cost change 1s->1s

There were no other plan changes in internal testing/pipelines.