ICW flake: Move gpsd and minirepro tests to run without autoanalyze (#17729)
I've seen some flakes with the gpsd and minirepro tests. In some of
these tests, we load specific stats and then verify that the correct
stats were loaded/match. However, if autoanalyze is enabled, we'll
sometimes get flakes if autoanalyze happened to run between when the
stats were loaded and when they're checked. To avoid this, move these
tests to not run with autoanalyze.

Example: in https://prod.ci.gpdb.pivotal.io/teams/main/pipelines/gpdb_main_without_asserts/jobs/icw_gporca_resgroup_rocky8/builds/751,
I noticed that we insert specific stats with a value of -1 for stadistinct, but in some flakes this is overwritten to 0.